refactor: 3-layer 전환 3단계 — leagueentry/common 모듈 Port 제거
- leagueentry: SummonerSeedRegistry Port/Impl 삭제, LeagueEntriesFetcher가 SummonerJpaRepository 직접 사용 (tier.name() 변환 인라인)
- common: MatchPayloadReaderImpl/MatchQueueEnqueuerImpl → 동명 클래스로 리네임 (@Component → @Service), 두 Port 인터페이스 삭제
- CollectMatchIdsRunner 등 컨슈머 import 갱신, 관련 테스트 모두 새 타입으로 정렬
