--- src/main/java/com/bestduo_BE/common/application/port/MatchPayloadReader.java
@@ -1,7 +1,0 @@
-package com.bestduo_BE.common.application.port;
-
-import com.bestduo_BE.common.infra.riot.dto.RiotMatchDto;
-
-public interface MatchPayloadReader {
-  RiotMatchDto read(String matchId);
-}
