+++ src/main/java/com/bestduo_BE/common/infra/persistence/MatchQueueEnqueuer.java
@@ -1,26 +1,23 @@
 package com.bestduo_BE.common.infra.persistence;
 
-import com.bestduo_BE.common.application.port.MatchQueueEnqueuer;
 import com.bestduo_BE.common.domain.model.Tier;
 import com.bestduo_BE.common.infra.persistence.entity.MatchQueue;
 import com.bestduo_BE.common.infra.persistence.repository.MatchQueueJpaRepository;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Component;
+import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-@Component
+@Service
 @RequiredArgsConstructor
-public class MatchQueueEnqueuerImpl implements MatchQueueEnqueuer {
+public class MatchQueueEnqueuer {
   private final MatchQueueJpaRepository repo;
 
-  @Override
   @Transactional
   public void enqueueAllIdempotent(List<String> matchIds, Tier tier, int priority, String patch) {
     for (String matchId : matchIds) {
-      if (repo.existsById(matchId)) continue; // ✅ 멱등
+      if (repo.existsById(matchId)) continue;
       repo.save(MatchQueue.newReady(matchId, tier, priority, patch));
     }
   }
-
 }
